[DataGrid] Improve selection with keyboard
